Revert "Cambios en los valores de los precios"

+++ src/BL/BL_Factura.java
@@ -18,7 +18,7 @@ public class BL_Factura {
     private Date fechaFactura;
     private boolean cancelada;
     private String tipoFactura;
-    private int montoAbonado;
+    private double montoAbonado;
     private ArrayList<BL_ProductoFactura> productosFactura  = new ArrayList<>();
     private BL_Cliente cliente;
     
@@ -65,11 +65,11 @@ public class BL_Factura {
         this.tipoFactura = tipoFactura;
     }
 
-    public int getMontoAbonado() {
+    public double getMontoAbonado() {
         return montoAbonado;
     }
 
-    public void setMontoAbonado(int montoAbonado) {
+    public void setMontoAbonado(double montoAbonado) {
         this.montoAbonado = montoAbonado;
     }
     
@@ -95,7 +95,7 @@ public class BL_Factura {
         return insertado;
     }
     
-    public void abonarAFactura(int monto){
+    public void abonarAFactura(double monto){
         new DAO_Factura().abonarAFactura(idFactura, monto);
     }
